requirements-custom.txt
sly
uvloop
orjson
psutil
lru_dict
distro
asyncpg
chess
humanize
fantasy_names
dateparser
discord.py-custom
aiowiki
sentry_sdk
redis[hiredis]
aiohttp[speedups]
tomli
